using open-ai base transformers and prettify
